automation/requirements.txt
# Python Dependencies for FairMediator Automation
# ALL FREE - No paid services required!
# Install with: pip install -r requirements.txt

# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# FREE Hugging Face Integration
huggingface-hub==0.19.4
transformers==4.36.0

# Web Scraping (FREE)
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Data Processing (FREE)
pandas==2.1.3
numpy==1.26.2

# MongoDB (FREE tier available)
pymongo==4.6.0

# Scheduling (FREE)
schedule==1.2.0

# Optional: Local inference (100% FREE, runs on your machine)
# torch==2.1.1  # Uncomment if you want to run models locally
# sentencepiece==0.1.99  # For tokenization
